Change conda guide URLs to relative
The current links to the conda guide (condaHPC.md) are full URL pointing to a different repo. This changes the URLs just to file names which will result in links pointing to the same repo (i.e., this repo). This fixes broken links if you don't have access to the other repo.